+++ gaios/Wallet/WalletViewController/Cells/TransactionCell/TransactionCell.swift
@@ -39,7 +39,7 @@ class TransactionCell: UITableViewCell {
         self.imgView.image = model.icon
 
         var txtCache = ""
-        let registry = WalletManager.current?.registry
+        let registry = WalletManager.current
         for (idx, amount) in model.amounts.enumerated() {
             let asset = registry?.info(for: amount.key)
             if let balance = Balance.fromSatoshi(amount.value, assetId: amount.key) {