+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### [Unreleased]
+
+### Changed
+- Improve Assets app cache Behavior
+
 ## [4.0.27] - 2024-05-03
 
 ### Added